Create DNASequence Classification-Kernel
# DNA sequences classification

Project for the course *Advanced Learning Models* (2017/2018) at Grenoble INP.

## Contraints

The project require the development of a classification system for DNA sequences. The choice was to create a kernel based SVM classifier with different possibilities for the kernels.
In this case the data are preprocessed to obtain the a spectral-like mapping considering different length sequences.
## Play Around

The code as it is should have the best performances but you can play around with different kernels or different mappings. Just enjoy

### Requirements

- Pandas
- cvxopt

## Authors

* **Moreno La Quatra** [MorenoLaQuatra](https://github.com/MorenoLaQuatra)